fix(form-builder): create FormControl instead of FormGroup when the type of a property is an object but no properties are provided to it.
BREAKING CHANGE: Until now we created FormGroup's when the type of a property was an object but no properties were provided to it. This lead to us defining properties as strings even when we wanted to use them as objects in custom fields. Additionally  we no longer support wrongly formatted object definitions: {type: 'object', items: {properties: {}}}. Only the valid format: {type: 'object', properties: {}} is now support.